Evangeli/requirements.txt
fastapi
uvicorn
pydantic
python-binance